fix compiler warning: implicit conversion shortens 64-bit value into a 32-bit value

+++ lib/curl_addrinfo.c
@@ -5,7 +5,7 @@
  *                            | (__| |_| |  _ <| |___
  *                             \___|\___/|_| \_\_____|
  *
- * Copyright (C) 1998 - 2008, Daniel Stenberg, <daniel@haxx.se>, et al.
+ * Copyright (C) 1998 - 2009, Daniel Stenberg, <daniel@haxx.se>, et al.
  *
  * This software is licensed as described in the file COPYING, which
  * you should have received as part of this distribution. The terms
@@ -370,7 +370,7 @@ Curl_ip2addr(int af, const void *inaddr, const char *hostname, int port)
   struct namebuff *buf;
   char  *addrentry;
   char  *hoststr;
-  int    addrsize;
+  size_t addrsize;
 
   DEBUGASSERT(inaddr && hostname);
